+++ docs/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Switch from pipfile into poetry and from setup.py into pyproject.toml
 - Decorator should not start an context if it's not needed.
+- Change settings of DatabasePlugin from a form of "db:dbname:url" into normal dict: {"databases": {"dbname": {"url": "x"}}}
+- Add recreate for DatabasePlugin and remove old recreate mechanism, including using "default url".
+- Removed old "driver" support for DatabasePlugin. Preparing code for CQRS instead, like "command" and "query" wrappers.
 
 ## 0.4.0 - Split Context Manager and Decorator for Configurator